Suppressed clippy warnings

+++ crates/pyxel-extension/src/variable_wrapper.rs
@@ -34,7 +34,7 @@ impl Colors {
     }
 
     pub fn from_list(&mut self, lst: Vec<Rgb8>) {
-        let copy_size = usize::min(self.list().len() as usize, lst.len());
+        let copy_size = usize::min(self.list().len(), lst.len());
         self.list_mut()[..copy_size].clone_from_slice(&lst[..copy_size]);
     }
